src/test/unit/com/tramchester/domain/ProvidesNotesTest.java
package com.tramchester.domain;

import org.joda.time.LocalDate;
import org.junit.Before;
import org.junit.Test;

import java.util.List;

import static org.junit.Assert.assertEquals;

public class ProvidesNotesTest {
    private String expected = "At the weekend your journey may be affected by improvement works." +
            "Please check <a href=\"http://www.metrolink.co.uk/pages/pni.aspx\">TFGM</a> for details.";
    private ProvidesNotes provider;

    @Before
    public void beforeEachTestRuns() {
        provider = new ProvidesNotes();
    }

    @Test
    public void shouldAddNotesForSaturdayJourney() {
        TramServiceDate queryDate = new TramServiceDate(LocalDate.parse("2016-10-29"));
        List<String> result = provider.createNotesFor(queryDate);

        assertEquals(1,result.size());
        assertEquals(expected,result.get(0));
    }

    @Test
    public void shouldAddNotesForSundayJourney() {
        TramServiceDate queryDate = new TramServiceDate(LocalDate.parse("2016-10-30"));
        List<String> result = provider.createNotesFor(queryDate);

        assertEquals(1,result.size());
        assertEquals(expected,result.get(0));
    }

    @Test
    public void shouldNotShowNotesOnOtherDay() {
        TramServiceDate queryDate = new TramServiceDate(LocalDate.parse("2016-10-31"));
        List<String> result = provider.createNotesFor(queryDate);

        assertEquals(0,result.size());
    }
}
